+++ src/webhooks/types/events/connection_activated.rs
@@ -1,0 +1,6 @@
+use serde::Deserialize;
+
+use crate::sso::Connection;
+
+#[derive(Debug, Deserialize)]
+pub struct ConnectionActivatedWebhook(pub Connection);
